@@ -1,3 +1,8 @@
-# crawl4ai/_version.py
-__version__ = "0.6.3"
+# crawl4ai/__version__.py
+
+# This is the version that will be used for stable releases
+__version__ = "0.7.6"
+
+# For nightly builds, this gets set during build process
+__nightly_version__ = None
 
