@@ -2,7 +2,7 @@ import os
 import json
 from crawl4ai.web_crawler import WebCrawler
 from crawl4ai.chunking_strategy import *
-from crawl4ai.extraction_strategy import *
+from crawl4ai import *
 from crawl4ai.crawler_strategy import *
 
 url = r"https://marketplace.visualstudio.com/items?itemName=Unclecode.groqopilot"
